pyproject.toml
[project]
name = "podcast-adblocker"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "feedparser>=6.0.11",
    "audioop-lts>=0.2.1",
    "google-cloud-speech>=2.32.0",
    "google-cloud-storage>=3.1.0",
    "google-genai>=1.13.0",
    "pydub>=0.25.1",
    "pyright>=1.1.400",
    "ruff>=0.11.8",
    "invoke>=2.2.0",
    "requests>=2.32.3",
]
